requirements.txt
asn1crypto==0.24.0
Click==7.0
cryptography==2.1.4
Django==1.11.22
django-bulma==0.5.6
django-jquery==3.1.0
djangorestframework==3.9.4
enum34==1.1.6
Flask==1.0.2
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.10.1
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
mpmath==1.1.0
numpy==1.16.3
opencv-python==4.1.0.25
pycrypto==2.6.1
pygobject==3.26.1
pytz==2019.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
sympy==1.4
virtualenv==16.6.1
Werkzeug==0.15.2
